+++ inst/testme/test-sourceFutureStartupScript.R
@@ -74,18 +74,31 @@ stopifnot(!exists("startup_script_2", envir = globalenv()))
 rm("startup_script_was_sourced", envir = globalenv())
 reset()
 
-## The environment variable may list several scripts, separated by ':' or ';'
-##
-## FIXME: sourceFutureStartupScript() splits R_FUTURE_STARTUP_SCRIPT on
-## '[:;]', which does not work on MS Windows, because the drive letter is
-## split off too, e.g. 'C:/path/.future.R' becomes 'C' and '/path/.future.R'.
-## Neither of those exist, so on MS Windows an absolute pathname is silently
-## ignored - also when only a single script is specified. Because of this,
-## this part is only tested on non-Windows platforms.
+## The environment variable may list several scripts. On MS Windows they are
+## separated by ';' only, because ':' is part of the drive letter of an
+## absolute pathname
+Sys.setenv(R_FUTURE_STARTUP_SCRIPT =
+             paste(file.path(path, "nope.R"), script,
+                   sep = .Platform$path.sep))
+res <- sourceFutureStartupScript()
+print(res)
+stopifnot(identical(res, script))
+rm("startup_script_was_sourced", envir = globalenv())
+reset()
+
+## A single, absolute pathname works too, which on MS Windows means the
+## drive letter must not be mistaken for a separator
+Sys.setenv(R_FUTURE_STARTUP_SCRIPT = script)
+res <- sourceFutureStartupScript()
+print(res)
+stopifnot(identical(res, script))
+rm("startup_script_was_sourced", envir = globalenv())
+reset()
+
+## On non-Windows platforms, ';' also works as a separator
 if (.Platform$OS.type != "windows") {
   Sys.setenv(R_FUTURE_STARTUP_SCRIPT =
-               paste(file.path(path, "nope.R"), script,
-                     sep = .Platform$path.sep))
+               paste(file.path(path, "nope.R"), script, sep = ";"))
   res <- sourceFutureStartupScript()
   print(res)
   stopifnot(identical(res, script))